services/ml/requirements.txt
ultralytics==8.2.0
torch
torchvision
torchaudio
opencv-python
pandas
matplotlib
PyYAML
tqdm
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
redis
rq
